mi_codigo_reflex/requirements.txt
httptools==0.6.1
lazy_loader==0.4
psycopg2-binary==2.9.11
python-dotenv==1.0.1
PyYAML==6.0.1
reflex==0.5.4
pydantic>=2.0
sqlmodel==0.0.14
